Create script.js
jQuery is used to target DOM elements and to add CSS classes that will be used to perform the class defined transitions to grow circle. Once complete the setTimeout function kicks in with .css transition settings to decrease the size or the circle to optimum.

Transition settings in the class bubble are 0.5 seconds per property to grow however 0.25 seconds per property here defined in the setTimeout function prior to adding the bubblepop2 class. Therefore the bubble grow is slower than the bubble shrink effect.